fix tests + clear warnings + refactor
