+++ src/tests/testPopupsManager.ts
@@ -1,5 +1,6 @@
-import {popupInstance, PopupManager} from '../popupManager';
+import {PopupManager} from '../popupManager';
 import { TestPopup  } from './testPopups';
+import {popupInstance} from '../popupsDef';
 
 export class TestPopupsManager extends PopupManager {
   public openTestPopup(dataHook: string, onClose?: () => void, content?: string): popupInstance {
